requirements.txt
Django>=4.2.0
python-dotenv>=1.0.0
requests>=2.31.0
networkx>=3.1
googlemaps>=4.10.0
python-decouple>=3.8
django-bootstrap5>=23.3
reportlab>=4.0.4
pillow>=10.0.0
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7
